requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py37_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
blas=2.11=openblas
blessings=1.7=py37hc8dfbb8_1001
blinker=1.4=py37_0
byted-dataloader=0.2.1=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37_1003
cheroot=8.3.0=pypi_0
click=7.1.1=py_0
colorama=0.4.3=py_0
conda=4.8.3=py37_0
conda-package-handling=1.6.0=py37h7b6447c_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
dbus=1.13.14=hb2f20db_0
dill=0.3.1.1=pypi_0
docopt=0.6.2=pypi_0
easydict=1.9=py_0
expat=2.2.6=he6710b0_0
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h86ecdb6_1001
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gast=0.2.2=py37_0
giflib=5.1.4=h14c3975_1
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
gnutls=3.6.5=h71b1129_1002
google-auth=1.13.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
gpustat=0.6.0=py_0
graphite2=1.3.13=h23475e2_0
grpcio=1.27.2=py37hf8bcb03_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=nompi_py37h513d04c_102
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.5=nompi_h3c11f04_1104
hdfs=2.5.8=pypi_0
icu=64.2=he1b5a44_1
idna=2.8=py37_0
imageio=2.8.0=pypi_0
imutils=0.5.3=py37_1
intel-openmp=2019.4=243
jaraco-functools=3.0.0=pypi_0
jasper=1.900.1=hd497a04_4
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37he6710b0_0
lame=3.100=h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=11_openblas
libcblas=3.8.0=11_openblas
libclang=9.0.1=default_hde54327_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
liblapack=3.8.0=11_openblas
liblapacke=3.8.0=11_openblas
libllvm9=9.0.1=he513fc3_1
libopenblas=0.3.6=h5a2b251_2
libopencv=4.2.0=py37_5
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libwebp=1.0.2=h576950b_1
libxcb=1.13=h1bed415_1
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
matplotlib-base=3.2.1=py37h30547a4_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37hc1659b7_1
mkl_random=1.1.0=py37hb3f55d8_0
more-itertools=8.2.0=pypi_0
multiprocess=0.70.9=pypi_0
ncurses=6.2=he6710b0_0
nettle=3.4.1=hbb512f6_0
ninja=1.8.2=py37h6bb024c_1
nspr=4.22=he6710b0_0
nss=3.46.1=hab99668_0
numpy=1.18.1=py37h94c655d_0
numpy-base=1.18.1=py37h2f8d375_1
nvidia-ml=7.352.0=py_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=4.2.0=py37_5
openh264=2.1.0=hd408876_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=3.1.0=py_0
p-tqdm=1.3.3=pypi_0
pandas=1.0.3=py37h0573a6f_0
pathos=0.2.5=pypi_0
pcre=8.43=he6710b0_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
pixman=0.38.0=h7b6447c_0
pox=0.2.7=pypi_0
ppft=1.6.6.1=pypi_0
protobuf=3.11.4=py37he6710b0_0
psutil=5.7.0=py37h7b6447c_0
py-opencv=4.2.0=py37h43977f1_5
py4j=0.10.7=py37_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycosat=0.6.3=py37h7b6447c_0
pycparser=2.19=py37_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pysocks=1.7.1=py37_0
pyspark=2.4.5=py_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.3.1=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2019.3=py_0
pyyaml=5.3.1=py37h7b6447c_0
qt=5.12.5=hd8c4c69_1
readline=7.0=h7b6447c_5
requests=2.22.0=py37_1
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
ruamel_yaml=0.15.87=py37h7b6447c_0
scikit-learn=0.22.1=py37h22eb022_0
scipy=1.4.1=py37habc2bb6_0
setuptools=45.2.0=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tensorboard=2.1.0=py3_0
tensorboardx=2.0=py_0
tensorflow=2.1.0=mkl_py37h80a91df_0
tensorflow-base=2.1.0=mkl_py37h6d63fb7_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
termcolor=1.1.0=py37_1
thrift=0.13.0=py37he6710b0_0
tk=8.6.10=hed695b0_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.44.1=py_0
urllib3=1.25.8=py37_0
web-py=0.51=pypi_0
werkzeug=1.0.1=py_0
wget=3.2=pypi_0
wheel=0.34.2=py37_0
wrapt=1.12.1=py37h7b6447c_1
x264=1!157.20191217=h7b6447c_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
